fix compile warning when passing a 32 bit type into an 8 bit type.

+++ Components/RTShaderSystem/include/OgreShaderParameter.h
@@ -589,7 +589,7 @@ public:
             return;
 
         mParamsPtr->_setRawAutoConstant(mPhysicalIndex, mAutoConstantType, data, mVariability,
-                                        mElementSize);
+            static_cast<uint8>(mElementSize));
     }
 
 protected:
